+++ packages/@expo/config-plugins/CHANGELOG.md
@@ -6,6 +6,8 @@
 
 ### 🎉 New features
 
+- Add Android `pathAdvancedPattern` intent filter support. ([#26116](https://github.com/expo/expo/issues/26116) by [@mvincentong](https://github.com/mvincentong))
+
 ### 🐛 Bug fixes
 
 ### 💡 Others